+++ extensions/xdebug/lib/XDebugExtension.php
@@ -18,10 +18,7 @@ use PhpBench\Executor\CompositeExecutor;
 use PhpBench\Extension\CoreExtension;
 use PhpBench\Extensions\XDebug\Command\Handler\OutputDirHandler;
 use PhpBench\Extensions\XDebug\Command\ProfileCommand;
-use PhpBench\Extensions\XDebug\Command\TraceCommand;
 use PhpBench\Extensions\XDebug\Executor\ProfileExecutor;
-use PhpBench\Extensions\XDebug\Executor\TraceExecutor;
-use PhpBench\Extensions\XDebug\Renderer\TraceRenderer;
 
 class XDebugExtension implements ExtensionInterface
 {
@@ -37,14 +34,6 @@ class XDebugExtension implements ExtensionInterface
         $container->register('xdebug.command.profile', function (Container $container) {
             return new ProfileCommand(
                 $container->get('console.command.handler.runner'),
-                $container->get('xdebug.command.handler.output_dir')
-            );
-        }, ['console.command' => []]);
-
-        $container->register('xdebug.command.trace', function (Container $container) {
-            return new TraceCommand(
-                $container->get('console.command.handler.runner'),
-                $container->get('xdebug.renderer.trace'),
                 $container->get('xdebug.command.handler.output_dir')
             );
         }, ['console.command' => []]);
@@ -65,25 +54,6 @@ class XDebugExtension implements ExtensionInterface
             ['benchmark_executor' => ['name' => 'xdebug_profile'],
         ]);
 
-        $container->register('xdebug.executor.xdebug_trace',
-            function (Container $container) {
-                return new CompositeExecutor(
-                    new TraceExecutor(
-                        $container->get(CoreExtension::SERVICE_EXECUTOR_BENCHMARK_MICROTIME)
-                    ),
-                    $container->get(CoreExtension::SERVICE_EXECUTOR_METHOD_REMOTE)
-                );
-            },
-            ['benchmark_executor' => ['name' => 'xdebug_trace'],
-        ]);
-
-        $container->register('xdebug.renderer.trace', function (Container $container) {
-            return new TraceRenderer(
-                $container->get('phpbench.formatter')
-            );
-        });
-
         $container->mergeParameter('executors', require_once(__DIR__ . '/config/executors.php'));
-        $container->mergeParameter('reports', require_once(__DIR__ . '/config/generators.php'));
     }
 }